requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
aiohttp==3.9.1
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
attrs==23.1.0
    # via aiohttp
cffi==1.16.0
    # via pynacl
discord-py[voice]==2.3.2
    # via -r .\requirements.in
ffmpeg-python==0.2.0
    # via -r .\requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
future==0.18.3
    # via ffmpeg-python
idna==3.6
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pycparser==2.21
    # via cffi
pynacl==1.5.0
    # via discord-py
pytube==15.0.0
    # via -r .\requirements.in
yarl==1.9.3
    # via aiohttp
youtube-dl==2021.12.17
    # via -r .\requirements.in
